+++ src/components/pages/index.test.jsx
@@ -19,6 +19,7 @@ import GithubLink from './logos/GithubLink';
 import RawContentContainer from './RawContentContainer';
 import MenuBar from './MenuBar';
 import PageContainer from './PageContainer';
+import Loading from './Loading';
 
 describe('Components snapshots', () => {
   it('renders the links correctly', () => {
@@ -36,6 +37,11 @@ describe('Components snapshots', () => {
     const rawContentContainerTree = renderer.create(<RawContentContainer />).toJSON();
 
     expect(rawContentContainerTree).toMatchSnapshot();
+  });
+  it('renders the Loading correctly', () => {
+    const tree = renderer.create(<Loading />).toJSON();
+
+    expect(tree).toMatchSnapshot();
   });
   it('renders the MenuBar correctly', () => {
     const tree = renderer.create(<Router><MenuBar routes={[]} /></Router>).toJSON();